% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{print_details}
\alias{print_details}
\title{Print a formatted version of the result details}
\usage{
print_details(x)
}
\arguments{
\item{x}{output of the \code{aziz.test()} function}
}
\description{
Print a formatted version of the result details
}
\seealso{
\code{\link{print_summary}}
}
