% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{print_summary}
\alias{print_summary}
\title{Print a formatted version of the results}
\usage{
print_summary(x)
}
\arguments{
\item{x}{output of the \code{aziz.test()} function}
}
\description{
Print a formatted version of the results for clarity
}
\seealso{
\code{\link{print_details}} for more info
}
