% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{reformat_results}
\alias{reformat_results}
\title{Reformats multiple results into one table}
\usage{
reformat_results(res_esa)
}
\arguments{
\item{res_esa}{listed outputs of multiple calls to \code{aziz.test()} on multiple variables}
}
\value{
A data.frame containing all results in an accessible presentation
}
\description{
If running multiple variables (and storing in a list),
this transform the list of results to one coherent data.frame
}
